+++ src/main/java/com/ipn/mx/SportConnect/repositorios/DeportivoRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DeportivoRepository extends JpaRepository<Deportivo, Integer> {
+public interface DeportivoRepository extends JpaRepository<Deportivo, Long> {
     //Todos los métodos CRUD
     @Query("SELECT d FROM Deportivo d JOIN FETCH d.encargado e WHERE e.rfcCurp = :rfcCurp")
     List<Deportivo> findDeportivosByEncargado(@Param("rfcCurp") String rfcCurp);
